# apps/core-api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/workshop/register": {
      "post": {
        "operationId": "WorkshopController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterIntakeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Workshop"
        ]
      }
    },
    "/api/workshop/orders": {
      "post": {
        "operationId": "WorkshopController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkshopOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkshopOrderResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Workshop"
        ]
      },
      "get": {
        "operationId": "WorkshopController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "sortField",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortDirection",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/WorkshopOrderResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Workshop"
        ]
      }
    },
    "/api/workshop/orders/{id}": {
      "get": {
        "operationId": "WorkshopController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkshopOrderResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Workshop"
        ]
      },
      "patch": {
        "operationId": "WorkshopController_updateOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkshopOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkshopOrderResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Workshop"
        ]
      }
    },
    "/api/workshop/orders/{id}/tasks": {
      "post": {
        "operationId": "WorkshopController_createTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkshopTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkshopTaskResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Workshop"
        ]
      }
    },
    "/api/workshop/orders/{id}/pick-parts": {
      "post": {
        "operationId": "WorkshopController_pickParts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PickWorkshopPartsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workshop parts pick transfer summary.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickWorkshopPartsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Workshop"
        ]
      }
    },
    "/api/workshop/orders/{orderId}/tasks/{taskId}": {
      "patch": {
        "operationId": "WorkshopController_updateTask",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkshopTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkshopOrderResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Workshop"
        ]
      },
      "delete": {
        "operationId": "WorkshopController_deleteTask",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workshop task deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkshopOrderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Task cannot be deleted because the order is invoiced or already has a linked invoice.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Workshop task or order was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Workshop"
        ]
      }
    },
    "/api/workshop/orders/{orderId}/tasks/{taskId}/line-items": {
      "patch": {
        "operationId": "WorkshopController_replaceTaskLineItems",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceWorkshopTaskLineItemsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkshopOrderResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Workshop"
        ]
      }
    },
    "/api/workshop/orders/{id}/create-invoice": {
      "post": {
        "operationId": "WorkshopController_createInvoiceFromOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Workshop"
        ]
      }
    },
    "/api/workshop/search": {
      "get": {
        "operationId": "WorkshopController_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Workshop"
        ]
      }
    },
    "/api/workshop/orders/{id}/pdf": {
      "post": {
        "operationId": "WorkshopController_generatePdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Workshop PDF generation status.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkshopPdfGenerationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Workshop"
        ]
      },
      "get": {
        "operationId": "WorkshopController_getPdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workshop PDF",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "tags": [
          "Workshop"
        ]
      }
    },
...
  },
  "info": {
    "title": "Auto Core Platform API",
    "description": "Generated OpenAPI spec for contract checks and client types.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "pageCount": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "page",
          "pageSize",
          "pageCount"
        ]
      },
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "RegisterIntakeDto": {
        "type": "object",
        "properties": {}
      },
      "CreateWorkshopOrderDto": {
        "type": "object",
        "properties": {}
      },
      "WorkshopCustomerSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "COMPANY"
            ]
          },
          "company_name": {
            "type": "string",
            "nullable": true
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "type",
          "first_name",
          "last_name"
        ]
      },
      "WorkshopVehicleSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "make": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "year": {
            "type": "number"
          },
          "vin": {
            "type": "string",
            "nullable": true
          },
          "plate": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "make",
          "model",
          "year"
        ]
      },
      "WorkshopTaskLineItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "LABOR",
              "PART"
            ]
          },
          "itemNo": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "qty": {
            "type": "number"
          },
          "unitPrice": {
            "type": "number"
          },
          "laborOperationId": {
            "type": "string",
            "nullable": true
          },
          "standardAw": {
            "type": "number",
            "nullable": true
          },
          "actualHours": {
            "type": "number",
            "nullable": true
          },
          "internalCostRate": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "id",
          "type",
          "itemNo",
          "description",
          "qty",
          "unitPrice"
        ]
      },
      "WorkshopTaskResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "WAITING_PARTS",
              "DONE"
            ]
          },
          "mechanic_notes": {
            "type": "string",
            "nullable": true
          },
          "mechanicNotes": {
            "type": "string",
            "nullable": true
          },
          "done": {
            "type": "boolean"
          },
          "line_items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkshopTaskLineItemResponseDto"
            }
          },
          "lineItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkshopTaskLineItemResponseDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "done",
          "lineItems",
          "createdAt",
          "updatedAt"
        ]
      },
      "WorkshopInvoiceSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "invoice_number": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id"
        ]
      },
      "WorkshopOrderResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "order_number": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "INTAKE",
              "IN_PROGRESS",
              "COMPLETED",
              "INVOICED"
            ]
          },
          "customer_id": {
            "type": "string"
          },
          "vehicle_id": {
            "type": "string"
          },
          "odometer": {
            "type": "number"
          },
          "fuel_level": {
            "type": "number"
          },
          "reported_issue": {
            "type": "string",
            "nullable": true
          },
          "reportedIssue": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "staging_location_id": {
            "type": "string",
            "nullable": true
          },
          "stagingLocationId": {
            "type": "string",
            "nullable": true
          },
          "customer": {
            "$ref": "#/components/schemas/WorkshopCustomerSummaryDto"
          },
          "vehicle": {
            "$ref": "#/components/schemas/WorkshopVehicleSummaryDto"
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkshopTaskResponseDto"
            }
          },
          "invoice": {
            "$ref": "#/components/schemas/WorkshopInvoiceSummaryDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "order_number",
          "status",
          "customer_id",
          "vehicle_id",
          "odometer",
          "fuel_level",
          "customer",
          "vehicle",
          "tasks",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateWorkshopOrderDto": {
        "type": "object",
        "properties": {}
      },
      "CreateWorkshopTaskDto": {
        "type": "object",
        "properties": {}
      },
      "PickWorkshopPartsLineDto": {
        "type": "object",
        "properties": {
          "workshopTaskLineItemId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "integer",
            "example": 4,
            "minimum": 1
          },
          "sourceLocationId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "workshopTaskLineItemId",
          "quantity"
        ]
      },
      "PickWorkshopPartsDto": {
        "type": "object",
        "properties": {
          "destinationLocationId": {
            "type": "string",
            "format": "uuid"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PickWorkshopPartsLineDto"
            }
          }
        },
        "required": [
          "destinationLocationId",
          "items"
        ]
      },
      "PickWorkshopPartAllocationResponseDto": {
        "type": "object",
        "properties": {
          "sourceLocationId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "number",
            "example": 2,
            "minimum": 1
          },
          "referenceId": {
            "type": "string"
          }
        },
        "required": [
          "sourceLocationId",
          "quantity",
          "referenceId"
        ]
      },
      "PickWorkshopPartMovedLineResponseDto": {
        "type": "object",
        "properties": {
          "workshopTaskLineItemId": {
            "type": "string",
            "format": "uuid"
          },
          "movedQuantity": {
            "type": "number",
            "example": 4,
            "minimum": 1
          },
          "allocations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PickWorkshopPartAllocationResponseDto"
            }
          }
        },
        "required": [
          "workshopTaskLineItemId",
          "movedQuantity",
          "allocations"
        ]
      },
      "PickWorkshopPartsResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "stagingLocationId": {
            "type": "string",
            "format": "uuid"
          },
          "transferGroupId": {
            "type": "string",
            "example": "WO-PICK-<order-id>-<timestamp>"
          },
          "movedLines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PickWorkshopPartMovedLineResponseDto"
            }
          }
        },
        "required": [
          "id",
          "stagingLocationId",
          "transferGroupId",
          "movedLines"
        ]
      },
      "UpdateWorkshopTaskDto": {
        "type": "object",
        "properties": {}
      },
      "ReplaceWorkshopTaskLineItemsDto": {
        "type": "object",
        "properties": {}
      },
      "WorkshopPdfGenerationResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "enqueued": {
            "type": "boolean"
          },
          "taskId": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "enqueued"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}